+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix a weird crash when a child has 0 effective height
+
 ## [0.1.5] - 2025-07-06
 
 - Delayed UI updates when loading a save state, fixed some graphical issue